requirements_blip_env.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
bitsandbytes==0.49.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
datasets==4.4.2
debugpy==1.8.19
decorator==5.2.1
dill==0.4.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.7.0
jupyter_core==5.9.1
MarkupSafe==2.1.5
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.18.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
sympy==1.13.1
tdqm==0.0.1
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
